invoke main() via __libc_start_main()
This commit rewritten the code for _start() stub.  It is necessary to
invoke __libc_start_main() so that libc can have a chance to initialize
the its internal data structure and get the chance to flush the buffered
stdout.

For example, if we compile the code with:

    $ qemu-arm -L/usr/arm-linux-gnueabihf ./amacc -o a.out tests/char.c

Then, run it with I/O redirection:

    $ qemu-arm -L/usr/arm-linux-gnueabihf ./a.out > out.txt

You will find that there was no output in out.txt before this commit.