fix incorrect result with nested for loop
Close #13
The for loop in previous version is parsed as:
// Init -> Cond -> Bz -> After -> Body -> Jmp
and then swap the After and Body chunk.
The swap does not consider the jump address in Body chunk and thus the
incorrect jump address causes unexpected behavior. This issue has been
reported at issue #13

In this patches for loop is parsed as:
Init -> Cond -> Bz to end -> Jmp to Body ->
After -> Jmp to Cond -> Body ->Jmp to After

It disadvantage at generating more Jmp instruction, which may decrease
overall performance. However, it is more convenient then recalculating
all jump address in Body and After chunk.