fix tests/char.c by removing -fsigned-char
This commit removes -fsigned-char from the Makefile because ARM ABI
says `char` should be unsigned by default, unless `-fsigned-char` is
specified.

We should remove this flag because amacc.c uses following trick to
detect whether signed char should be used for the target environment:

    neg_char = 255;
    neg_int = neg_char;

BTW, assuming host environment (for the compiler) is equal to the target
environment (for the generated output executable) is not the right thing
to do.  We would probably like to introduce `-fsigned-char` to amacc as
well.