Integrate Abstract Syntax Tree + backend code generator (#50)
This patch introduces the following enhancements:
1. Abstract syntax tree (AST) creation
2. Backend code generator function: gen()
3. Standard ordering of function parameters on stack

Specifically speaking, it adds an abstract syntax tree, constant folding, and
a split front-end/back-end design. With an AST, the compiler is not limited to generating
code on the fly with parsing. This capability allows function parameter code to be
emitted and pushed on the stack in the proper right-to-left order.

With a modular code generator, new targets are straightforwardly supported such as
native Arm machine code and ELF output.

close #37