Allow amacc to bootstrap compile.
Before this fix, you could only bootstrap compile amacc once.
This fix allows a chain of bootstrap compiles as far out as
you want to go.

This involved a longstanding bug having to do with
literal constant and/or Enum subtractions, which is now fixed.

It also involved a bug having to do with the type system,
for which a workaround was implemented.