Squint -- a peephole optimizer for amacc
This pull request is a starting point for discussion on how to integrate
Squint most effectively with amacc.  It may best be done with separate
repos that are tied through modules, but I don't know.

I added a -p option to amacc.c that must currently be used with the
amacc -o option to produce an ELF file that can be optimized with
scripts/peep.  In order for peep to work, a version of squint must
be compiled, e.g. gcc -o squint squint.c .

Example:

% gcc -o amacc-gcc amacc.c -ldl
% ./amacc-gcc -p -o amacc amacc.c
% scripts/disasm amacc > amacc.s
% gcc -o squint squint.c
% scripts/peep amacc
executable file amacc was optimized.
% scripts/disasm amacc > amacc.s.opt
% sdiff -W -d -w 180 amacc.s amacc.s.opt

I have more advanced versions of this that has even beat gcc -O1 for fib.c
and possibly other things, but I suspect you will not like my coding style,
so we will have to discuss.