Simplify peephole frame variable to register transformation
I traced back a problem to a use-def analysis across a basic block boundary.
This is not a problem in my private version of Squint which is more complete.
I "dubded down" the algorithm, so it is no longer faster than gcc -O0, but
on the other hand, it is not much slower considering how much dumber the
algorithm is.