Allow use of array index with struct pointers (#88)
Using an array operator '[ ]' is currently supported for use with (certain) pointer
variables in array expressions, but not always. The goal of this patch is to be able
to use the array operator '[ ]' in expressions, working as expected under every
possible circumstance. Note that having '[ ]' work for declarations also needs to
be improved.

Close #83 

Co-authored-by: HPCguy <Anonymous>