Address reviewer issues for Suint.c pull request.
There is a lot here to keep track of.  It would be nice
to address any further issues in a subsequent pull request.
A very small Squint follow-up pull requests is planned to
improve/finalize squint testing and to add an option to
configure squint as a shared library in addition to current
standalone mode.